requirements.txt
allure-python-commons
appium-python-client==3.1.0
selenium==4.18.1
pytest==8.0.0
pytest-bdd==7.0.0
allure-pytest==2.13.2
pyyaml==6.0.1
pytest-rerunfailures
ruff
